Fireblaze

Use fireblaze to manage authentications.

Public fields

session

A valid Shiny session.

Active bindings

signed_in

Read the signed in user.

signed_up

Read the signed in user.

Methods

Public methods

Fireblaze$new()

Fireblaze$sign_out()

Fireblaze$get_signed_in()

Fireblaze$get_signed_up()

Fireblaze$set_language_code()

Fireblaze$clone()

Method new()

Usage

Fireblaze$new(
  persistence = c("session", "none", "local"),
  config_path = "fireblaze.rds",
  session = shiny::getDefaultReactiveDomain()
)

Arguments

persistence

How the auth should persit: none, the user has to sign in at every visit, session will only persist in current tab, local persist even when window is closed.

config_path

Path to the configuration file as created by create_config.

session

A valid shiny session.

Details

Initialise Fireblaze

Method sign_out()

Usage

Fireblaze$sign_out()

Details

Signs user out

Method get_signed_in()

Usage

Fireblaze$get_signed_in()

Details

Signed in user details triggered when auth states changes

Method get_signed_up()

Usage

Fireblaze$get_signed_up()

Details

Signed up user details triggered explicitely by user (e.g.: on click)

Method set_language_code()

Usage

Fireblaze$set_language_code(code)

Arguments

code

iso639-1 language code.

Details

Set language code for auth provider

Method clone()

The objects of this class are cloneable with this method.

Usage

Fireblaze$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.